+++ Viewer/ImageDisplay.cpp
@@ -437,6 +437,16 @@ bool Viewer::ImageDisplay::filterHistogramEqualization()
     return true;
 }
 
+bool Viewer::ImageDisplay::canRotate()
+{
+    return true;
+}
+
+bool Viewer::ImageDisplay::canZoom()
+{
+    return true;
+}
+
 void Viewer::ImageDisplay::updateZoomCaption()
 {
     const QSize imgSize = m_loadedImage.size();