+++ Viewer/ImageDisplay.h
@@ -59,6 +59,9 @@ public:
     bool filterContrastStretch();
     bool filterHistogramEqualization();
 
+    bool canRotate() override;
+    bool canZoom() override;
+
 public Q_SLOTS:
     void zoomIn() override;
     void zoomOut() override;