+++ Viewer/VLCDisplay.h
@@ -33,6 +33,8 @@ public:
     QImage screenShoot() override;
     void relativeSeek(int msec) override;
     void restart() override;
+    bool canRotate() override;
+    bool canZoom() override;
 
 public Q_SLOTS:
     void updateInterface();